requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
groq>=0.4.0
python-dotenv>=1.0.0
discord.py>=2.3.0
python-telegram-bot>=21.0.0
psutil>=5.9.0
pyperclip>=1.8.0
mss>=9.0.0
pyfiglet>=1.0.0
tavily-python>=0.3.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
